fix delete for New App members;

+++ scripts/deleteuser.php
@@ -23,33 +23,38 @@ $user->setEmail($uinfo['email']);
 try
 {
     $validationService->email($user);
-    $authService->User($user->getEmail());
 }
-catch (\CognitoException | UserAccountNotExist | EmailValidationException $exception)
+catch (EmailValidationException $exception) // user email/pwd validation errors thrown by validationService
 {
     echo json_encode(array('status' => 'failed', 'reason' => $exception->getMessage()));
     exit();
 }
 
-try
-{
-    $authService->DeleteUser($uinfo['email']);
-}
-catch (CognitoException $exception)
-{
-    echo json_encode(array('status' => 'failed', 'reason' => $exception->getMessage()));
-    exit();
-}
 
 $action = 'delete';
-sendMail($uinfo['email'], $uinfo['firstname'], "EpiCore Unsubscription Notification", $action, $uinfo['maillist_id']);
 
 $result = UserInfo::deleteMaillist($uid);
 $status = $result[0];
 $message = $result[1];
 
 if($status == 'success') {
-    print json_encode(array('status' => 'success'));
+
+    sendMail($uinfo['email'], $uinfo['firstname'], "EpiCore Unsubscription Notification", $action, $uinfo['maillist_id']);
+    //delete on Cognito
+    try
+    {
+        if($message === 'deletion of maillist & fetp success.') {
+            $authService->User($user->getEmail()); 
+            $authService->DeleteUser($uinfo['email']);
+        } 
+        print json_encode(array('status' => 'success'));
+    }
+    catch (\CognitoException | UserAccountNotExist $exception)
+    {
+        echo json_encode(array('status' => 'failed', 'reason' => $exception->getMessage()));
+        exit();
+    }
+
 } else {
     print json_encode(array('status' => 'failed', 'reason' => $message));
 }
